requirements.txt
Django~=4.1
djangorestframework
psycopg2
psycopg2-binary
requests
PyYAML